requirements.txt
requests
selenium
Pillow
ttkbootstrap
